fix: view 컨트롤러와 일반 컨트롤러의 응답 구분
