+++ src/main/java/gift/exception/GlobalExceptionHandler.java
@@ -1,28 +1,26 @@
 package gift.exception;
 
+import gift.controller.ProductController;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-@ControllerAdvice
+@RestControllerAdvice(assignableTypes = ProductController.class)
 public class GlobalExceptionHandler {
 
     @ExceptionHandler(ProductNotExistException.class)
-    public ModelAndView handleProductNotExist() {
-
-        Map<String, String> model = new HashMap<>();
-        model.put("errorMessage", "상품이 존재하지 않습니다.");
-
-        return new ModelAndView("error/product-not-found", model);
+    public ResponseEntity<ErrorResponse> handleProductNotExist(ProductNotExistException ex) {
+        return ResponseEntity
+                .status(HttpStatus.NOT_FOUND)
+                .body(new ErrorResponse(ex.getMessage()));
     }
-
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ResponseEntity<Map<String, List<String>>> handleValidationExceptions(MethodArgumentNotValidException ex) {
@@ -37,4 +35,7 @@ public class GlobalExceptionHandler {
 
         return ResponseEntity.badRequest().body(errors);
     }
+
+    public record ErrorResponse(String message) {}
+
 }
